# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API Handle Products.",
        "title": "Certified Tech Developer - Dentistas",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/dentist": {
            "post": {
                "description": "Add a dentist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dentists"
                ],
                "summary": "Add a dentist",
                "parameters": [
                    {
                        "description": "Dentist",
                        "name": "dentist",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dentist.Dentist"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "TOKEN",
                        "in": "header"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dentist.Dentist"
                        }
                    }
                }
            }
        },
        "/dentist/{id}": {
            "get": {
                "description": "Get dentist by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dentists"
                ],
                "summary": "Get dentist by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dentist.Dentist"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a dentist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dentists"
                ],
                "summary": "Update a dentist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dentist",
                        "name": "turno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dentist.Dentist"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "TOKEN",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dentist.Dentist"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a dentist",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dentists"
                ],
                "summary": "Delete a dentist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "TOKEN",
                        "in": "header"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "dentist deleted",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/turno": {
            "get": {
                "description": "Get turnos by Paciente ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "turnos"
                ],
                "summary": "Get turnos by Paciente ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DNI",
                        "name": "dni",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/turno.Turno"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a turno",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "turnos"
                ],
                "summary": "Add a turno",
                "parameters": [
                    {
                        "description": "Turno",
                        "name": "turno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/turno.Turno"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "TOKEN",
                        "in": "header"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/turno.Turno"
                        }
                    }
                }
            }
        },
        "/turno/{id}": {
            "get": {
                "description": "Get turno by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "turnos"
                ],
                "summary": "Get turno by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/turno.Turno"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a turno",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "turnos"
                ],
                "summary": "Update a turno",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Turno",
                        "name": "turno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/turno.Turno"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "TOKEN",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/turno.Turno"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a turno",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "turnos"
                ],
                "summary": "Delete a turno",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "TOKEN",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "turno deleted",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update a turno with a patch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "turnos"
                ],
                "summary": "Update a turno with a patch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "TOKEN",
                        "in": "header"
                    },
                    {
                        "description": "Turno",
                        "name": "turno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/turno.Turno"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/turno.Turno"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dentist.Dentist": {
            "type": "object",
            "required": [
                "apellido",
                "matricula",
                "nombre"
            ],
            "properties": {
                "apellido": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "matricula": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
        "turno.Turno": {
            "type": "object",
            "properties": {
                "dentistaId": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                },
                "fechaHora": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "pacienteDni": {
                    "type": "string"
                }
            }
        }
    }
}